[reporting] update pdfmake node module (#121585)
* [Reporting] Upgrade pdfmake module

* change pdf maker to integration test

(cherry picked from commit 7e7319b00797e9e7324dd416d7a50048caf1cade)
