+++ package.json
@@ -306,7 +306,7 @@
     "p-map": "^4.0.0",
     "p-retry": "^4.2.0",
     "papaparse": "^5.2.0",
-    "pdfmake": "^0.1.65",
+    "pdfmake": "^0.2.4",
     "peggy": "^1.2.0",
     "pluralize": "3.1.0",
     "pngjs": "^3.4.0",
@@ -600,7 +600,7 @@
     "@types/ora": "^1.3.5",
     "@types/papaparse": "^5.0.3",
     "@types/parse-link-header": "^1.0.0",
-    "@types/pdfmake": "^0.1.15",
+    "@types/pdfmake": "^0.1.19",
     "@types/pegjs": "^0.10.1",
     "@types/pngjs": "^3.4.0",
     "@types/prettier": "^2.3.2",
